Process results.

+++ scripts/read_logs_zip_results.py
@@ -60,12 +60,13 @@ if __name__ == '__main__':
             new_name = log_file.replace(log_basename, new_basename)
             result_basename = os.path.basename(result_file)
             shutil.move(log_file, new_name)
-            remove = raw_input('Remove results? y/[n]')
-            if remove.lower() == 'y':
-                if os.path.isfile(result_file):
-                    os.remove(result_file)
-                else:
-                    print("Couldn't remove " + result_file)
+            if result_file:
+                remove = raw_input('Remove results? y/[n]')
+                if remove.lower() == 'y':
+                    if os.path.isfile(result_file):
+                        os.remove(result_file)
+                    else:
+                        print("Couldn't remove " + result_file)
         else:
             print('Not finished yet')
